pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "termchat"
version = "0.1.0"
description = "CLI chat tool using OpenRouter with Tavily search integration"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
include = ["termchat*"]

[project.scripts]
termchat = "termchat.main:main"
